C++ switch
Write a C++ program that reads a number (a float) and a character from
the console. Depending on the character (see the options on the table
below), print the resulting math function (column 2 of the table)
applied on the number input by the user. Any other letter should not
affect the program.